+++ src/vec3.h
@@ -6,7 +6,7 @@
 /*   By: joaoteix <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/03 13:19:19 by joaoteix          #+#    #+#             */
-/*   Updated: 2024/04/03 01:07:01 by jcat             ###   ########.fr       */
+/*   Updated: 2024/04/05 04:49:50 by jcat             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,18 +17,18 @@
 #include <stdbool.h>
 
 typedef struct s_vec3 {
-	double	x;
-	double	y;
-	double	z;
+	float	x;
+	float	y;
+	float	z;
 }	t_vec3;
 
-double	length(t_vec3 *v);
+float	length(t_vec3 *v);
 bool	is_normal(t_vec3* v);
 
 t_vec3	vec3(const float x, const float y, const float z);
 t_vec3	v3sum(const t_vec3 a, const t_vec3 b);
 t_vec3	v3sub(const t_vec3 a, const t_vec3 b);
-double	v3dot(const t_vec3 u, const t_vec3 v);
+float	v3dot(const t_vec3 u, const t_vec3 v);
 t_vec3	v3cross(const t_vec3 a, const t_vec3 b);
 t_vec3	v3unit(const t_vec3 a);
 t_vec3	v3scalef(const t_vec3 v, const double f);
